feat: added x-client-version header, raw block subscription types
